Small Improvements for LegendItem
Added 3 new parameters LegendItem:
horSpacing: Specifies the spacing between the line symbol and the label.
verSpacing: Specifies the spacing between individual entries of the
legend vertically. (Can also be negative to have them really close)
box: Specifies if the Legend should will be drawn with a rectangle
around it.

This made it possible to change the resize behavior by just using
layout.setVerticalSpacing and layout.setVerticalSpacing.

Changed the removeItem function to accept either name or the originally
added Item. I found the latter more useful, especially if several items
happen to have the same name.

Changed the rows in which addItem inserts a new item. Now the number
increases all the time. This was needed because there where some
overwriting issues when using several times removeItem and addItem on
the same legend (in random order). The GraphicsGridLayout does not care
about empty rows, so this seems to work best.